renamed messages controller to calculator controller
