+++ conf/app.routes
@@ -1,3 +1,3 @@
 # microservice specific routes
 
-GET        /hello-world             controllers.MessageController.helloWorld
+GET        /hello-world             controllers.CalculatorController.helloWorld
