+++ test/controllers/CalculatorControllerSpec.scala
@@ -16,15 +16,26 @@
 
 package controllers
 
-import uk.gov.hmrc.play.microservice.controller.BaseController
-import play.api.mvc._
-import scala.concurrent.Future
+import play.api.http.Status
+import play.api.test.FakeRequest
+import uk.gov.hmrc.play.test.{UnitSpec, WithFakeApplication}
+import org.jsoup._
 
-object MessageController extends MessageController
 
-trait MessageController extends BaseController {
+class CalculatorControllerSpec extends UnitSpec with WithFakeApplication {
 
-  def helloWorld() = Action.async { implicit request =>
-    Future.successful(Ok("Hello world"))
+  "GET /capital-gains-calculator/hello-world" should {
+
+    val fakeRequest = FakeRequest("GET", "/capital-gains-calculator/hello-world")
+    val result = CalculatorController.helloWorld()(fakeRequest)
+
+    "return 200" in {
+      status(result) shouldBe Status.OK
+    }
+
+    "return the text `Hello world`" in {
+      val jsoupDoc = Jsoup.parse(bodyOf(result))
+      jsoupDoc.body.text shouldEqual "Hello world"
+    }
   }
 }
